Added eos-setup-registry
Add tools and services to resize and encrypt(key exchange)
the read-write container registry of suse-eos during boot.
suse-eos uses a split registry with one part in the read-only(squashfs)
region and one part on an extra partition that is writable.
The tools and services here target the read-write part. The
read-only setup of the regsitry is done in the suse-eos
image description at build time of the image